+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.2.8
 ++++++
 * `vm create --storage-sku`: can now specify the storage account sku for managed os and data disks separately.
+* `sig image-version`: Version names now consistently specified by  `--image-version -e`. `--image-version-name` deprecated.
 
 2.2.7
 ++++++